Make Nano S separate from Nano X
