+++ nix/fhs.nix
@@ -15,7 +15,7 @@ in pkgs.buildFHSUserEnv {
     xorg.libXScrnSaver xorg.libXcomposite xorg.libXcursor xorg.libXdamage
     xorg.libXext xorg.libXfixes xorg.libXi xorg.libXrandr xorg.libXrender
     xorg.libXtst xorg.libxcb xorg.xcbutilkeysyms zlib zsh
-    gnumake libtinfo5 glibc_multi.dev (python2.withPackages (ps: [ps.pillow]))
+    gnumake libtinfo5 glibc_multi.dev (python3.withPackages (ps: [ps.pillow ps.ledgerblue]))
   ];
   inherit runScript;
 }
