+++ src/apdu_baking.c
@@ -6,7 +6,7 @@
 #include "os_cx.h"
 #include "protocol.h"
 #include "to_string.h"
-#include "ui_prompt.h"
+#include "ui.h"
 
 #include <string.h>
