--- src/ui_menu.h
@@ -1,3 +1,0 @@
-#pragma once
-
-void main_menu(void);
